+++ src/ui/widgets/album_artist_list_widget.py
@@ -14,7 +14,7 @@ This is distinct from the Artists tab which groups by the *track-level* artist
 """
 from __future__ import annotations
 
-from typing import Iterable
+from collections.abc import Iterable
 
 from PySide6.QtCore import QItemSelectionModel, QModelIndex, Qt, Signal
 from PySide6.QtGui import QStandardItem, QStandardItemModel
@@ -167,7 +167,7 @@ class AlbumArtistListWidget(QWidget):
 
     def set_ui_scale(self, scale: float) -> None:
         self._ui_scale = max(0.85, min(1.5, float(scale or 1.0)))
-        self.table.verticalHeader().setDefaultSectionSize(int(round(30 * self._ui_scale)))
+        self.table.verticalHeader().setDefaultSectionSize(round(30 * self._ui_scale))
         self.album_browser.set_ui_scale(self._ui_scale)
 
     def apply_current_palette(self) -> None:
@@ -411,7 +411,7 @@ class AlbumArtistListWidget(QWidget):
         try:
             counts = get_album_artist_letter_counts(self.app_state.db, self._search)
             self.alpha_index.refresh(counts)
-        except Exception:
+        except Exception:  # noqa: BLE001, S110
             pass
 
     def set_ignore_articles(self, ignore: bool) -> None:
